replace formatData func
