+++ src/app/services/table-data.service.ts
@@ -3,7 +3,8 @@ import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { tableData } from '../interfaces/fixed-table-data-interfaces';
-import { jsonData } from '../interfaces/back-data-interfaces';
+import { formatData } from '../../assets/formatData';
+
 
 @Injectable({
   providedIn: 'root',
@@ -14,19 +15,6 @@ export class TableDataService {
 
   public receiveData(sectionName: string): Observable<tableData> {
     return this.http.get(this.url).pipe(
-      map( data => this.formatData(data[sectionName])));
-  }
-
-  private formatData(limb: jsonData): tableData {
-    let result = Object.assign(limb);
-
-    Object.keys(result).map(
-      element => {
-        typeof result[element] === 'object' ?
-          result[element] = this.formatData(result[element]) :
-          result = Object.values(result);
-      });
-
-    return result;
+      map( data => formatData(data[sectionName])));
   }
 }
