+++ src/assets/formatData.ts
@@ -1,0 +1,15 @@
+import { jsonData } from '../app/interfaces/back-data-interfaces';
+import { tableData } from '../app/interfaces/fixed-table-data-interfaces';
+
+export function formatData(limb: jsonData): tableData {
+  let result = Object.assign(limb);
+
+  Object.keys(result).map(
+    element => {
+      typeof result[element] === 'object' ?
+        result[element] = formatData(result[element]) :
+        result = Object.values(result);
+    });
+
+  return result;
+}
